requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
google-generativeai
faiss-cpu
langchain_community
langchain_google_genai
matplotlib
plotly
google-cloud-translate